+++ source/Nuke.Common/Tools/ReSharper/ReSharperTasks.cs
@@ -17,27 +17,26 @@ partial class ReSharperTasks
 {
     public const string ReSharperPluginLatest = "latest";
 
-    protected override ToolOptions PreProcess(ToolOptions options)
+    protected override T PreProcess<T>(T options)
     {
-        var resharperOptions = ((ReSharperSettingsBase)options)
-            .SetProcessToolPath(GetToolPath());
-        if (resharperOptions.Plugins is not { Count: > 0 })
-            return resharperOptions;
+        options = options.SetProcessToolPath(GetToolPath(options));
+        if (options is not ReSharperSettingsBase { Plugins.Count: > 0, Plugins: var plugins })
+            return options;
 
-        var wave = GetWave(resharperOptions).NotNull("wave != null");
+        var wave = GetWave().NotNull("wave != null");
         var shadowDirectory = NukeBuild.TemporaryDirectory / $"ReSharper-{wave}-{GetHash()}";
 
-        ((AbsolutePath)resharperOptions.ProcessToolPath).Parent.Copy(
+        ((AbsolutePath)options.ProcessToolPath).Parent.Copy(
             target: shadowDirectory,
             policy: ExistsPolicy.MergeAndOverwriteIfNewer);
 
         DownloadPlugins();
 
-        return resharperOptions
-            .SetProcessToolPath(shadowDirectory / Path.GetFileName(resharperOptions.ProcessToolPath));
+        return options
+            .SetProcessToolPath(shadowDirectory / Path.GetFileName(options.ProcessToolPath));
 
-        string GetWave(ReSharperSettingsBase toolSettings) =>
-            Directory.GetParent(toolSettings.ProcessToolPath)
+        string GetWave() =>
+            Directory.GetParent(options.ProcessToolPath)
                 .DescendantsAndSelf(x => x.Parent)
                 .Select(x => Path.Combine(x.FullName, "jetbrains.resharper.globaltools.nuspec"))
                 .Where(File.Exists)
@@ -46,13 +45,13 @@ partial class ReSharperTasks
                 .SingleOrDefault();
 
         string GetHash() =>
-            resharperOptions.ProcessToolPath
-                .Concat(resharperOptions.Plugins.Select(x => x.Key + x.Value))
+            options.ProcessToolPath
+                .Concat(plugins.Select(x => x.Key + x.Value))
                 .OrderBy(x => x).JoinComma()
                 .GetMD5Hash()[..4];
 
         void DownloadPlugins() =>
-            resharperOptions.Plugins
+            plugins
                 .Select(x => (Plugin: x.Key, Version: x.Value == ReSharperPluginLatest ? null : x.Value))
                 .ForEach(x => HttpTasks.HttpDownloadFile(
                     $"http://resharper-plugins.jetbrains.com/dotnet/api/v2/curated-feeds/Wave_v{wave}.0/Packages(Id='{x.Plugin}',Version='{x.Version}')/Download",