+++ source/Nuke.Components/ISignPackages.cs
@@ -11,7 +11,6 @@ using Nuke.Common;
 using Nuke.Common.CI.AppVeyor;
 using Nuke.Common.IO;
 using Nuke.Common.Utilities.Collections;
-using static Nuke.Common.IO.FileSystemTasks;
 using static Nuke.Common.Tools.SignPath.SignPathTasks;
 
 namespace Nuke.Components;